Started correctly distributing the functionality between 'br-locale-switcher' and 'br-locale-utility'.
