+++ brjs-sdk/sdk/libs/javascript/br-locale-switcher/test-unit/jsTestDriver.conf
@@ -1,0 +1,8 @@
+server: http://localhost:4224
+basepath: .
+
+load:
+  - bundles/bundle.js
+
+test:
+  - tests/**.js
