+++ brjs-sdk/sdk/libs/javascript/br-services/src/br/services/locale/BRLocaleService.js
@@ -53,7 +53,7 @@ BRLocaleService.prototype.getLocale = function() {
 	var localeCookieValue = this.localeUtility.getCookie( localeCookieName );
 	var browserLocales = this.localeUtility.getBrowserAcceptedLocales();
 	var appLocales = this.appMetaService.getLocales();
-	return LocaleUtility.getActiveLocale( localeCookieValue, browserLocales, appLocales );
+	return LocaleUtility.getActiveLocaleGiven( localeCookieValue, browserLocales, appLocales );
 };
 
 /**
